+++ presentation/src/main/java/com/fakedevelopers/presentation/ui/productEditor/ProductEditorDto.kt
@@ -27,10 +27,3 @@ data class ProductEditorDto(
         categoryId = categoryId
     )
 }
-
-data class ProductCategoryDto(
-    val categoryId: Long,
-    val categoryName: String,
-    val parentCategoryId: Long,
-    val subCategories: List<ProductCategoryDto>
-)
